Module.target() now uses craftr.utils.dis.get_assigned_name()
The taget name is now optional to not break backwards compatibility.
Compiler modules have also been updated to have the target name as
an optional argument and added value checks for required parameters.

The platform module was also updated as it didn't take the change
of the platform.windows.name value into account and errored on Windows.