relative target names with dots can now be references with double-colons
For example, the main Craftr module is "project" and there is a target
in the module "project.libs" called "autotools". If you would pass the
target "libs.autotools" to Craftr, it would assume that an absolute target
name was specified due to the period in it. Previously, you were required
to write the full target name like "project.libs.autotools". With this
commit, you can also pass "libs:autotools" instead.